Set DST automatically
Hopefully a fix to the DST that should work in the future by it self. Sets the
beginDST if the current date is "summertime".

Also added a trigger for the webhook. This is not done anymore by the
SolarMeter project since it is now reading the values directly from
the inverter.